Moving job to src/jobs
